move more validators code back to validators.tsx
